requirements.txt
streamlit
mysql.connector